photomaker-side/requirements.txt
replicate
gradio
python-dotenv
boto3